pc-1318 masquer les offres d'activations de types thing
- par defaut a la creation du compte user, on cree automatiquement une offre de reservation de type numerique
- il faut masquer cette offre afin que l'user n'est pas acces a la contremarque
