+++ src/selectors/selectBookings.js
@@ -3,6 +3,15 @@ import moment from 'moment'
 import difference from 'lodash.difference'
 import { createSelector } from 'reselect'
 
+const filterBookingsActivationOfTypeThing = bookingobj => {
+  const offer = get(bookingobj, 'stock.resolvedOffer')
+  const offerType = get(offer, 'eventOrThing.type')
+  const isActivationType = offerType === 'EventType.ACTIVATION'
+  if (!isActivationType) return true
+  const offerThingId = get(offer, 'thingId')
+  return !offerThingId
+}
+
 export const selectBookings = createSelector(
   state => state.data.bookings,
   allBookings => allBookings
@@ -10,22 +19,29 @@ export const selectBookings = createSelector(
 
 export const selectSoonBookings = createSelector(
   // select all bookings
-  state => state.data.bookings,
+  selectBookings,
   allBookings => {
     const twoDaysFromNow = moment().subtract(2, 'days')
-    const filtered = allBookings.filter(booking => {
-      const date = get(booking, 'stock.eventOccurrence.beginningDatetime')
-      if (!date) return false
-      return moment(date).isSameOrBefore(twoDaysFromNow)
-    })
+    const filtered = allBookings
+      .filter(booking => {
+        const date = get(booking, 'stock.eventOccurrence.beginningDatetime')
+        if (!date) return false
+        return moment(date).isSameOrBefore(twoDaysFromNow)
+      })
+      .filter(filterBookingsActivationOfTypeThing)
     return filtered
   }
 )
 
 export const selectOtherBookings = createSelector(
-  state => state.data.bookings,
+  selectBookings,
   selectSoonBookings,
-  (allBookings, soonBookings) => difference(allBookings, soonBookings)
+  (allBookings, soonBookings) => {
+    const filtered = difference(allBookings, soonBookings).filter(
+      filterBookingsActivationOfTypeThing
+    )
+    return filtered
+  }
 )
 
 export const selectBookingById = createSelector(
